backend(models): Remade migrations, models and enums in process
Due to a laravel convention by creating tables in plural (ej: create_users_table), I had to delete the past migrations and rewrite their names to plural, this for the model to be capable of recognize the table its apointing to in the database. Models are in process, still referencing the relationships. Factories aren't done yet, so the project could have problems / warnings related to it.